refactor: extract truststore encoding to internal package
